services/wallet-service/requirements.txt
fastapi
uvicorn[standard]
motor
python-jose[cryptography]
cryptography
pydantic
python-dotenv
